+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.0 - 2024-01-11
+### Added
+- Bulk translate action (requires [Multi Translator Plugin](https://plugins.craftcms.com/multi-translator))
+### Fixed
+- import from csv with languages mapped from other languages
+
 ## 3.0.1 - 2023-12-28
 ### Fixed
 - copy matrix elements in Craft 5
@@ -11,6 +17,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 3.0.0 - 2023-12-28
 ### Added
 - Craft 5 support
+
+## 2.5.0 - 2024-01-11
+### Added
+- Bulk translate action (requires [Multi Translator Plugin](https://plugins.craftcms.com/multi-translator))
+### Fixed
+- import from csv with languages mapped from other languages
 
 ## 2.4.0 - 2023-12-18
 ### Added